requirements.txt
Django==5.1.6
djangorestframework==3.15.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-filter==24.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1 